+++ src/libtrx/game/objects/general/save_crystal.c
@@ -1,10 +1,9 @@
+#include "config.h"
 #include "game/game_flow.h"
+#include "game/input.h"
+#include "game/lara.h"
 #include "game/objects/common.h"
-
-#include <libtrx/config.h>
-#include <libtrx/game/input.h>
-#include <libtrx/game/lara.h>
-#include <libtrx/game/rooms.h>
+#include "game/rooms.h"
 
 static const OBJECT_BOUNDS m_SaveCrystal_Bounds = {
     .shift = {