+++ src/tr1/game/objects/common.c
@@ -162,36 +162,3 @@ void Object_DrawPickupItem(const ITEM *const item)
 
     Matrix_Pop();
 }
-
-void Object_SetMeshReflective(
-    const OBJECT_ID obj_id, const int32_t mesh_idx, const bool enabled)
-{
-    const OBJECT *const obj = Object_Get(obj_id);
-    if (!obj->loaded) {
-        return;
-    }
-
-    OBJECT_MESH *const mesh = Object_GetMesh(obj->mesh_idx + mesh_idx);
-    mesh->enable_reflections = enabled;
-    for (int32_t i = 0; i < mesh->num_tex_face4s; i++) {
-        mesh->tex_face4s[i].enable_reflections = enabled;
-    }
-    for (int32_t i = 0; i < mesh->num_tex_face3s; i++) {
-        mesh->tex_face3s[i].enable_reflections = enabled;
-    }
-    for (int32_t i = 0; i < mesh->num_flat_face4s; i++) {
-        mesh->flat_face4s[i].enable_reflections = enabled;
-    }
-    for (int32_t i = 0; i < mesh->num_flat_face3s; i++) {
-        mesh->flat_face3s[i].enable_reflections = enabled;
-    }
-    Output_DispatchObjectMeshUpdate(obj->mesh_idx + mesh_idx);
-}
-
-void Object_SetReflective(const OBJECT_ID obj_id, const bool enabled)
-{
-    const OBJECT *const obj = Object_Get(obj_id);
-    for (int32_t i = 0; i < obj->mesh_count; i++) {
-        Object_SetMeshReflective(obj_id, i, enabled);
-    }
-}